GtkSharp.BlazorWebView.csproj: cleanup code
