--- src/BlazorWebView/src/GtkSharp.BlazorWebView/BlazorWebViewOptions.cs
@@ -1,13 +1,0 @@
-using System;
-using System.IO;
-
-namespace GtkSharp.BlazorWebKit;
-
-public class BlazorWebViewOptions
-{
-
-	public Type RootComponent { get; init; }
-
-	public string HostPath { get; init; } = Path.Combine("wwwroot", "index.html");
-
-}
