--- src/BlazorWebView/src/GtkSharp.BlazorWebView/ServiceCollectionExtensions.cs
@@ -1,13 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-
-namespace GtkSharp.BlazorWebKit;
-
-public static class ServiceCollectionExtensions
-{
-    public static IServiceCollection AddBlazorWebViewOptions(this IServiceCollection services, BlazorWebViewOptions options)
-    {
-        return services
-            .AddBlazorWebView()
-            .AddSingleton<BlazorWebViewOptions>(options);
-    }
-}
